messaging_app/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
idna==3.10
parameterized==0.9.0
pillow==11.2.1
requests==2.32.3
sqlparse==0.5.3
typing==3.7.4.3
typing_extensions==4.13.2
urllib3==2.4.0
